Staging view: display conflict state information
Show more conflict details in the unstaged viewer of the staging view.

Propagate the StageState telling which stages exist in the index from
the IndexDiff up to the StagingEntry. Use it there to display what kind
of conflict there is: delete-modify, modify-delete, both modified, or
both added different content. Add a second column to the unstaged viewer
(conflicts are always unstaged) to show an abbreviation and a fuller
message on hover.

Add two new icons for modify-delete and delete-modify conflicts and
use them in the decorator.

Change-Id: Ia50796a61a8bbdd8ac9016c8f50c9711f699a407
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>