Ensure nested git repositories are picked up
EGit already detected .git resources inside newly connected Eclipse
projects and touched them to force a ResourceChangeEvent containing
them in the delta so that GitProjectData could connect the submodules.

However, that didn't always work. First, the corresponding code was
sometimes not called (for instance, when auto-sharing a project), and
second, it didn't work for nested repositories. Touching a folder does
not include it in the next resource delta.

Instead of touching ".git" resources and relying on a resource change
event collect the resources and pass them to GitProjectData directly.

Bug: 571984
Change-Id: I097ee672a70540e8cebb4e8492260fa8ce715c57
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>