Ensure nested git repositories are picked up
EGit already detected .git resources inside newly connected Eclipse
projects and touched them to force a ResourceChangeEvent containing
them in the delta so that GitDataProvider could connect the submodules.

However, that didn't always work. First, the corresponding code was
sometimes not called (for instance, when auto-sharing a project), and
second, it didn't work for nested repositories. Touching a folder does
not include it in the next resource delta.

Looks like the Eclipse resource tree is similar in that respect to git
and only handles files as expected.

Instead of touching ".git" folders, touch the HEAD file inside, if it
exists. If not, the directory isn't a git-directory anyway.

Bug: 571984
Change-Id: I097ee672a70540e8cebb4e8492260fa8ce715c57
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>