Provide a GitInfo accessor interface as API
Simplify accessing some git information for git managed items such
as IResources or IFileRevisions. Provide a simple GitInfo accessor
interface that can be obtained by third-party bundles via the
adapter mechanism, as in

  Adapters.adapt(someObject, GitInfo.class)

where someObject can be an IResource, IFileRevision, GitBlobStorage,
GitRemoteResource, or even a FileRevisionTypedElement.

This first version provides access to the JGit Repository and to the
git path of the item. With this, a third-party bundle can construct
the absolute file system path by appending the git path to the path
of repository.getWorkTree().

The idea is that third-party bundles should not have to use
EGit-internal APIs like RepositoryMapping or the RepositoryCache.

Change-Id: I89070bc83d56a81b3661e925ce0f5e23d4acaf11
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>