GitInfo API: access to commit ID and index state
Extend the GitInfo API to also provide information about where the
underlying git item is: in the working tree, in the index, or in a
commit. Add a getSource() method, and a getCommitId() method returning
the commit id if the item is from a commit.

Further extend the GitInfo API to provide information about the git
state of an item. Move the IResourceState interface from EGit UI to
EGit core, make it public and rename it to GitItemState. Add a method
GitInfo.getGitState() to get an instance. This gives third-party
bundles a stable API for safe and efficient access to EGit's
IndexDiffData. Otherwise clients would have to read and traverse the
whole DirCache from the repository to determine such info themselves.

Bug: 573222
Change-Id: I3f7bc7ad0330511c7aad705ee12216bcc4350e7f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>