[merge] Record in index whether conflicting file has been updated
When the merge input preference is set to "Last HEAD (unmerged)", EGit
uses the index stage 2 ('ours') as input. This is the version as it
was, without pre-merged content.

If the user then edits this version and saves, the new contents are
written to the working tree.

Previously, opening the merge editor then again on the same (not yet
resolved, i.e., still unstaged) conflict file would again show the
'ours' version, without the previously saved edits. For the user, his
previous edits appear lost even though they were saved.

Change GitMergeEditorInput to record in the git index whether the file
had been saved. The "last modified time" of a conflict stage has no
meaning in the git index, and is normally set to zero both in C git
and in JGit. EGit now changes the "last modified time" of the first
conflict stage to the current time upon saving, and before opening
the 'ours' side checks that timestamp and takes the working tree file
if the timestamp is not zero. That way the merge editor re-opens with
the last saved content including the user's edits.

Bug: 548813
Change-Id: I619f5cacea3ec4687f2023c99d8dd6b63f2ada1e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>