[merge editor] Fix base for cherry-picks
A cherry-pick does not use the common ancestor as base but a parent
of the cherry-picked commit. If the cherry-picked commit has a single
parent, use that; otherwise use stage 1 from the index with a general
label.

(If a merge commit is cherry-picked, git does not record the chosen
mainline, and there is no way after the fact to figure it out.)

Bug: 482154
Change-Id: Iec0e4c1516f2da6746ac3de53b071798449020a4
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>