[merge] Use EFS-linked IResources for non-workspace files
LocationEditableRevision caused JDT to log exceptions and to not
initialize the merge viewer fully. Fix this by creating a hidden
resource linked to the non-workspace file instead. These hidden
resources are all created in a hidden project, and they are cleaned
up when the merge editor is disposed. In case some of them remain
all the same, which can happen when the IDE shuts down, EGit core
removes any linked resources it finds in that hidden project on
startup.

JDT uses similar mechanisms for external folders, and so does the
Compare UI's CompareWithOtherResourceDialog.

Bug: 573232
Change-Id: If7c76b9480fc6dc14e888c7a21440146d37a8ad1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>