[merge] Hide changes between ancestor and current for cherry-pick
By default hide these changes. The ancestor in a cherry-pick is a
parent commit of 'theirs' (the commit that was cherry-picked); the
differences between that parent and the current version are simply
not interesting and even misleading.

Add an action to the content merge viewer's toolbar to allow the user
to toggle this setting for three-way comparisons.

Bug: 441149
Change-Id: I154cbecd445ef4481a1288c87c0e6e9cf498651f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>