Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Build Successful 
> 
> https://ci.eclipse.org/egit/job/egit-stable/171/ : SUCCESS

Not good. This build packaged the JGit m2 JARs into the update site. It also ran the build against JGit m2. And that now causes the trouble in https://git.eclipse.org/r/c/simrel/org.eclipse.simrel.build/+/181081 .

I don't see how this could have happened. The JGit build appears to have run and completed before this job started... though it's hard to tell with all the different time zones used by Gerrit, Jenkins, and maven. But probably it'd be wise to tighten the checks in the build scripts and require exactly matching suffixes for the -m and -r builds.

In any case it looks like this job should be retriggered.

Patch-set: 1
CC: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>