Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > How does one "un-publish" something from Nexus?
> 
> If [1] is correct only a Nexus admin could do this.
> 
> [1] https://stackoverflow.com/questions/40173808/something-like-mvn-undeploy-to-remove-artifacts-from-nexus

yes, release artefacts are supposed to be final.

At SAP we prevent this scenario by pushing release builds first to a Nexus staging repository and only when validation of all release artefacts was successful we release the staging repository.
Otherwise the staging repo is dropped which can be done by developers.

This staging feature is only available in the commercial Nexus Pro version [2].

[2] https://help.sonatype.com/repomanager3/staging

Patch-set: 1