Update patch set 1
Patch Set 1:

Another idea for the EGit builds might be to not rely on a tag in the upstream repo but enumerate the versions published at Nexus and take the latest matching one from there. Then it doesn't matter at all when the tag is pushed, or if there is a tag at all. Would have to look how one enumerates versions via some Nexus API, though. I presume it has some endpoint for this that we could call via curl.

Patch-set: 1