Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I tried to use this series to resolve the conflicts created when cherry-picking I76fb554f7bc09d61afb33f01b3b7fa3ad3d20c64
> onto master.
> 
> Though I couldn't understand how to use it.
> The lines showing how a hunk in the left pane fits to the right pane don't move when scrolling the source code. This is very confusing.

I cannot reproduce this. Scrolling behavior is the same as without this series.

But even with the fixed ancestor this still produces unusable comparisons for PackFile, no matter what one chooses as input for the left side. "Working tree merged to 'ours'" is best, but still not correctly shows the diffs as recorded in the working tree file. Hiding changes between left and ancestor also seems to hide a bit too much (with input as "Working tree (pre-merged by git)" it doesn't show the conflict in the imports). That's a bit worrying, and might be a bug in the compare framework. Or perhaps I'm not setting the right flags. Will investigate.

Perhaps Markus was right in bug 482154 and one should simply do a two-way comparison between "workspace resolved to 'ours'" and "workspace resolved to 'theirs'" for cherry-picks. The EFS would give us a way to implement that...

Disappointing. It worked so well with the conflicts from bug 482154.

Patch-set: 6