Update patch set 7
Patch Set 7:

> For PackFile, the results are not ideal no matter what input is chosen.

The result is also not ideal if left = "working tree pre-merged to ours" and right = "working tree pre-merged to theirs". Although that gives indeed a diff that matches the git conflict markers exactly, that diff isn't useful either. It would include also the ObjectLoader get(...) and the resolve() methods, which also should not be added to the master version of PackFile at all. It would apply and be syntactically correct, but would fail compilation.

With left = "working tree pre-merged to ours", right = stage 3 from index (theirs), and changes only between left and ancestor suppressed, I get a conflict in the imports and and an incoming addition of just the exists() method, which the Eclipse diff machinery would insert in a place where the result would be syntactically invalid. I do actually consider this a better comparison in this case; I see immediately that something is badly wrong. In this case, "wrong" means the versions differ so much that applying the change from one to the other is simply not possible.

Patch-set: 7