Update patch set 7
Patch Set 7:

Moreover, the results of doing the cherry-pick with command-line git 2.18.0 and with JGit differ.

Command-line git produces exactly what is shown in the EGit merge editor with this change when left = "working tree merged to 'ours'" and right = stage 3 (theirs), and changes between left and ancestor are hidden.

JGit records a different conflict in the working tree file.

Patch-set: 7