[compare] Refactor GitCompareEditorInput
GitCompareEditorInput is used only for comparing two commits, but it
also contained code that is supposed to deal with comparing the working
tree with a commit, or with comparisons involving the index.

Refactor this so that it makes use of the common comparison editor
things factored out to AbstractGitMergeEditorInput. Reduce the code for
now to deal only with comparisons between two commits. Improve the
labeling of the left and right sides to use the same labels as in the
GitCompareFileRevisionEditorInput.

Rename the abstract superclass to AbstractGitCompareEditorInput.

Change-Id: Ibdb71cc4d9449894c217c5c7eb81fab5e613aab2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>