[compare] Comparing the working tree with a commit
Re-add code to GitCompareEditorInput for comparing the working tree
against a commit. This functionality had existed previously, but had
been unused and was therefore removed in the refactoring in the parent
commit.

Now it's added again, and used in the repositories view: if a single
ref (branch, tag, or additional ref) is selected in the repositories
view, the CompareCommand now changes its label to "Compare with
Working Tree", and opens the compare editor with the comparison between
the working tree and the selected ref.

More uses can be added in future changes.

This kind of comparison does not replace the synchronize commands. The
synchronization support works only for IResources (i.e., files and
folders in the Eclipse workspace), but has elaborate support for
"model-based" comparisons. OTOH, this new "Compare with Working Tree"
command also works well for files not in the Eclipse workspace, but
cannot do model-based comparisons.

Change-Id: Iacc85443ab7e10762a27ec33dd668d47b0a56e2b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>