[releng] Force --release and use BREE to run tests
Make the EGit build work if code is compiled with a Java > 1.8.
This prepares for moving the BREE to Java 11: when we do that, we
will need to be able to build nightly with and against Java 11, but
the stable branch against Java 1.8.

The downside is that for a local maven build, you'll need a
~/.m2/toolchains.xml file providing JavaSE-1.8 and JavaSE-11.

The CI pods do provide such toolchain definitions for maven.

Once this change is submitted and merged to master the EGit CI
pipelines will be changed to no longer fix the Java version at
1.8.

Change-Id: I5f4c3ea944861d1c96dbbf8e4c2849de50b25219
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>