Update patch set 1
Patch Set 1:

OK, so the next version is 5.13. Fine by me.

We should really try to set up the maven builds such that they can work when maven is executed on Java 11 and still generate code for Java 8 and run the tests on a Java 8 JDK. We'll need that when we want to go to 6.0 and Java 11 (or otherwise we'll need to maintain different CI builds, which would be a shame.)

I got that working using maven toolchains (which are available in CI) for EGit, and also for JGit if javac is used. With the JGit ecj build I'm currently having problems, see bug 574181. (Didn't look at egit-github yet.) The downside of using toolchains is that one needs a $HOME/.m2/toolchains.xml even for local builds.

Patch-set: 1
CC: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>