Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> OK, so the next version is 5.13. Fine by me.

I don't want to dictate that, if we have breaking changes we want to get out in the next release we can still bump the version to 6.0. Though it's important to bump the version on master to avoid confusion by building the same version as on stable-5.12.

> We should really try to set up the maven builds such that they can work when maven is executed on Java 11 and still generate code for Java 8 and run the tests on a Java 8 JDK. We'll need that when we want to go to 6.0 and Java 11 (or otherwise we'll need to maintain different CI builds, which would be a shame.)

yes I agree

> I got that working using maven toolchains (which are available in CI) for EGit, and also for JGit if javac is used. With the JGit ecj build I'm currently having problems, see bug 574181. (Didn't look at egit-github yet.) The downside of using toolchains is that one needs a $HOME/.m2/toolchains.xml even for local builds.

AFAIR we added ecj to get the same warnings like we get in Eclipse.
As a workaround we can configure the maven build to use javac until we found the cause of the problem with ecj.

I am not sure what's the future of egit-github looking at Mylyn's current status.

Patch-set: 1