Update patch set 2
Patch Set 2:

Setup to run maven with either Java 8 or Java 11 to build EGit. Tests are run on Java 8 in either case; the BreeSmokeTest will fail if a compilation with Java 11 by mistake references things that exist only in a Java 11 library.

The new profile is necessary only to include the <release> tag conditionally because it is not allowed on Java 1.8. After the second release that uses Java 11 as BREE we could require Java 11 unconditionally for running the EGit maven build, and then we'd include <release>11</release> directly
in the main tycho-compiler-plugin configuration.

Build 244 used a Java 8 to compile, build 245 used a Java 11 (it used https://git.eclipse.org/r/c/egit/egit-pipelines/+/181929).

This means we can build EGit with either Java version, and the BREE is only in the MANIFEST.MF files and in the top-level pom.xml, so it can be changed between branches and the same pipeline scripts can be used.

If we do this, users will need a local $HOME/.m2/toolchains.xml as shown in bug 574181. I'd update the Contributor's Guide with the necessary info.

To test locally, create such a toolchains.xml, then run mvn with JAVA_HOME set to either a Java 8 or a Java 11 installation.

Patch-set: 2
Reviewer: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>