Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> can we update to tycho 2.x with this change ?

I think so. Tycho 2.x requires running the maven build with Java 11. Right now the builds still pin the Java version at 8. Once this is submitted and merged to master, I'll remove that pinning. Then it would be possible to use tycho 2.3.0. Personally I find it a little strange to require a Java 11 for building while the BREE is still Java 8, but we can do so.

Patch-set: 3