[index diff] Don't wait for the workspace lock
Updating an IndexDiffCacheEntry scheduled jobs that waited for the
workspace lock. This appears to be a remnant of olden times when EGit
used a WorkingTreeIterator that operated on the Eclipse resource tree.
But the ContainerTreeIterator and AdaptableFileTreeIterator were
removed in commit 983c7408 in 2015; since then, EGit operates on files
directly.

Remove waiting for the workspace lock.

Bug: 542597
Change-Id: I2d389f32bf24ce4a69e48f4102ac1ae1a2107b58
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>