Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I *believe* one nice side effect was that build and egit didn't run in parallel, at least egit waited - and this avoids extra disk IO while build is still running, so build could finish without interruptions.

Whether that's a _nice_ effect is debatable. It means the UI will update only once the build is done. I'm trying to avoid that.

Why should recomputing an index diff in a system job interrupt the build? It doesn't modify the file system, and anyway access the files directly, not through the Eclipse resource tree.

But I'm fully aware that I may not see everything here. This waiting for the workspace lock, later changed to use a scheduling rule encompassing the projects of a repo, is old code.

One possible problem might be builds creating/modifying non-ignored files. But that should cause resource changed events later on, and we'd update the index diff then again.

Patch-set: 1