[compare][history] Compare commit against working tree
Make "Compare with Working Tree" in the git history page work also
when the input is not filtered to a single file. The comparison is
still restricted by the current history page input, but now also
works if the input is several files, or one or more directories, or
the whole repository. Rename detection is not done in these cases.

Change-Id: Ib3c2fee79836de5765fdbde03b560927f73d628b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>