[compare] Disable "Link with Editor" for multi-file comparisons
The "Link with Editor" feature doesn't work well for comparisons of
multiple files. Leave only the "Show In" functionality (available
in the global "Navigate" menu) enabled. Ensure that it also works for
the folder nodes.

Getting the "Show In" sub-menu into the context menu of the
DiffTreeViewer is left for a follow-up commit.

Bug: 574710
Change-Id: I6252a7ff825d2b06b9290fb768c3bcd70bdddb37
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>