[history] Enable "Focus on this File" for bare repositories
Handle relative paths in a HistoryPageInput as git paths, and construct
such an input in CommitFileDiffViewer.

Also make "Show in->History" work in the compare editor (in the file
tree viewer top left): let FileDiff implement GitInfo, handle GitInfo
as input in GitHistoryPage, and in the commit editors produce such
GitInfo selections as input if no IResource or java.io.File can be
determined.

Bug: 574757
Change-Id: I48606ce5252c29a79b2f5f4afe5a2f5d26934850
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>