[history] Keep the target commit in view
Fix the logic to avoid "jumping" in the history view. Setting the top
index might scroll the just selected commit out of view. So swap the
logic: first set the top index, then select the target commit.

Also improve the handling if we have neither a known top commit nor a
target commit to select. In that case the table just kept its index,
but that might show now something completely unrelated, which is highly
confusing. Jump to the top in that case.

Bug: 574923
Change-Id: I4bd4dac2278cf99782347f84513d0581666a1218
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>