[blame] Follow in history only if linking with selection enabled
When navigating inside an editor showing revision information, EGit
updates the history view to show the commit corresponding to the
current text position. Do so only if "Link with Editor and Selection"
is activated for the history view.

Unfortunately there is no API to determine whether linking is enabled.
I've chosen to query the corresponding preference from the Team UI
plug-in.

Bug: 572886
Change-Id: Ied12d17b5eeeb73dd6391e66b1c064b33375e084
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>