Fix HistoryPageInput.equals()
Comparisons between BlameHistoryPageInput and HistoryPageInput
mistakenly could yield "true", and then navigating in a file with
revisions being shown would not follow commits in the history view
even if "Link with Editor and Selection" was enabled in the history
view.

Avoid creating duplicate history navigation entries: if the view
already shows the history of the file, just refresh the view and
select the commit.

Change-Id: I4fbe8e2da38ff2e0d31cb34f9075056e749768d7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>