Update patch set 2
Patch Set 2:

It seems if the selected commit (e.g. by selecting an old remote tracking branch) is too large, depending on the configured preference "History > Maximum number of commits to show", the history view jumps to the top probably since the history view starts to load the next chunk of older commits. Maybe we can improve the behavior so that it always jump to the correct commit. 

I tried this with a clone of jgit and then clicking remote tracking branches starting from the newest ones going back in the history (origin/master, origin/next, origin/stable-5.12 etc.). With the default setting max. number of commits of 10000 the history view follows the selection until I select stable-4.11. I guess at this point the history view decides to load additional commits and the selection jumps to the top. This doesn't happen when the max no of commits is set to 100000.

Patch-set: 2
Reviewer: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>