Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> It seems if the selected commit (e.g. by selecting an old remote tracking branch) is too large, depending on the configured preference "History > Maximum number of commits to show", the history view jumps to the top probably since the history view starts to load the next chunk of older commits. Maybe we can improve the behavior so that it always jump to the correct commit. 
> 
> I tried this with a clone of jgit and then clicking remote tracking branches starting from the newest ones going back in the history (origin/master, origin/next, origin/stable-5.12 etc.). With the default setting max. number of commits of 10000 the history view follows the selection until I select stable-4.11. I guess at this point the history view decides to load additional commits and the selection jumps to the top. This doesn't happen when the max no of commits is set to 100000.

I can kind of reproduce this but with a limit of 1000. In my first attempt it worked until stable-4.9; from 4.8 on nothing happens anymore. Then at 4.7 it jumps to the top. Afterwards it is broken. Going back to 4.10 works, but then 4.9 isn't selected even though I can see it in the history view. I think we are losing the commitToShow if it's not in the first batch, i.e., in the first (limit) commits.

But this bug also occurs without these changes. When I tried again without this change, it stopped working at 4.11, and the table kept the current index and showed me some completely unrelated commits. (Top commit was 4f83544e50, which is younger than 5.5.1.) I prefer jumping to the top. The remaining bug (commitToSelect lost if not loaded in first batch) will need to be handled separately.

Patch-set: 2