Update patch set 1
Patch Set 1:

This failure is bizarre:

  02:02:39  Errors: 
  02:02:39    GitCloneWizardHttpsTest.setup:34 » IO Invalid keystore format

Worst is: it reproduces locally on my machine with JDK 1.8_191.

Seems to be completely unrelated to this change, and I have no idea where to start looking for what might be wrong in the environment. The parent change used the same target platform and worked. JDK version doesn't seem to be the cause, surely jdk/latest on CI is newer than my old 1.8_191.

Patch-set: 1