Update patch set 2
Patch Set 2:

> where do we select the JDK used for the build ?

In the Egit builds we don't. We use whatever JDK is active to run maven (by default that's a Java 11 JDK on the Eclipse CI), and compile with ECJ using --release 8. For running tests, we use the toolchain defined by the BREE in the manifest (JavaSE-1.8) and rely on CI providing a maven toolchain definition. The log says it uses

  16:09:31  14:09:31 [INFO] Toolchain in tycho-surefire-plugin: JDK[/opt/tools/java/oracle/jdk-8/latest]

But apparent $PATH is set such that plain "keytool" is the keytool of a newer JDK.

Patch-set: 2