CommitSelectionDialog: restrict to single selection
The CommitSelectionDialog is supposed to support selecting a single
commit. Restrict the CommitGraphTable to not allow multi-selections.

Change the interface of CommitGraphTable to use style bits instead of
multiple booleans for feature selection. Style bits were chosen over
an enum because they are a better match with precedences in SWT or in
JFace.

Change-Id: I85accd37991f9d1c6cf2578a0ace6bd7325fb1ec
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>