CommitSelectionDialog: restrict to single selection
The CommitSelectionDialog is supposed to support selecting a single
commit. Restrict the CommitGraphTable to not allow multi-selections.

Change the interface of CommitGraphTable to use style bits instead of
multiple booleans for feature selection. Style bits were chosen over
an enum because they are a better match with precedences in SWT or in
JFace.

Change-Id: I165cd3b707fd6f48eddee10c777e0fd5fe9a4898
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>