[blame] Update revision information when HEAD changes
React on changes in HEAD of the repository a file is in and update the
revision annotations if the editor is visible. If it isn't queue the
update to occur once the editor becomes visible.

Remove installed listeners when revision annotations are hidden or the
editor is closed.

To avoid repeatedly installing listeners, store flags on the revision
ruler's control.

Bug: 433073
Change-Id: Ic2a46225a760290c81cc7c0738132e03ab9e9f10
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>