[blame] Update revision information when HEAD changes
React on changes in HEAD of the repository a file is in and update the
revision annotations if the editor is visible. If it isn't queue the
update to occur once the editor becomes visible.

Remove installed listeners when revision annotations are hidden or the
editor is closed.

To avoid repeatedly installing listeners, store flags on the revision
ruler's control.

Bug: 433073
Change-Id: I3c6ce6089eade2f551c476f7a29778f55ca7c6f5
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>