[blame] Fix creating an enriched hover control
Pass the last hovered line number via the BlameRevision to ensure the
sticky hover uses the same line number as the initial non-sticky hover
even if the mouse moved in the meantime. Add the status text to the
initial unfocused hover. When made sticky, fill the empty toolbar, for
now just with two buttons for opening the commit and for showing it in
the history view. More actions can be added in later changes.

Outwit the framework when a hover is made sticky via F2. In that case
the framework imposes a fairly small maximum size on the hover. But
revision hovers frequently can be large and thus would get resized to
a much smaller size when F2 is used. To prevent that make the size
of the initial hover known to the sticky hover so that it can ensure
it gets at least the same size.

Bug: 358232
Change-Id: I8b9280227aa188ba843a30549650552125aa3fc3
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>