[compare] Compare two arbitrary DiffNodes
Add a "Compare with Each Other" command to the GitDiffTreeViewer in
the Compare Editor. It enables the user to compare two arbitrary
files listed in the directory structure viewer, which is useful in
particular to deal with renamed files that are reported as an addition
and a deletion.

Platform/Compare already has a "Compare with Each Other" command for
resources, but it is implemented as an Action (object contribution). If
that were a Command, EGit could have re-purposed it by simply providing
a new handler for it.

EGit registers the new Action as a handler for a new command defined in
plugin.xml, so it accessible via Quick Access (Cmd/Ctrl-3), and users
can define a key binding for it. The action updates its own enablement
on selection changes in the GitDiffTreeViewer. The handler is only
active when the Compare Editor containing the viewer is active.

Change-Id: Ib10df32cc31f9c4852973520034e0cf67a9a6ac7
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>