Move the EGitCredentialsProvider to EGit core
This enables the EGit core to register a default credentials provider
when the core bundle is activated, which ensures a default provider is
always set.

If that CredentialsProvider needs a UI to interact with the user, it
calls a new CredentialsUI OSGi service. EGit UI publishes such a service
with the default ranking; other bundles could override this with their
own UI by providing a service with a higher ranking.

The EGitSecureStore is also made available as an OSGi service via a new
CredentialsStore service interface. As with the CredentialsProvider UI,
a third-party bundle could provide its own CredentialsStore by
publishing a service with a higher ranking.

Change-Id: I74d4f7e5f7d7d0a4af9f45c520ff67961b9653dd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>