Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I noticed that the toolbar isn't shown immediately, only when moving the mouse pointer inside the pop-up window the toolbar appears.

Or when you hit F2. The pop-up is enriched when it is made sticky, which is when the mouse pointer enters it or you hit F2. That's standard behavior provided by the framework, as in the javadoc pop-ups in the Java editor.

> Is it intended that the toolbar is at the bottom ? Took me a while to find it, first since it wasn't shown immediately and then I didn't expect it at the bottom.

Yes. The location of the toolbar is chosen by the framework.

> Finding the drop-down menu on the hunk header isn't obvious.

That's unchanged in this change; it's the normal HyperlinkDetector of the DiffViewer. When we added this HyperlinkDetector in the DiffViewer, we got this functionality for free also in the blame hovers.

Patch-set: 1