Don't refresh on working tree changes for non-EGit repositories
EGit refreshes its IndexDiffCache when JGit reports changes in working
trees. But this may also get triggered when third-party code uses
JGit directly to work with repositories. In such cases, refresh only
if the repository is known to EGit, i.e., is present in its
RepositoryCache.

Otherwise, there might be a new IndexDiffCacheEntry created without
repository in the RepositoryCache, which may result in a non-functional
IndexDiffCacheEntry that doesn't update on resource changes anymore.

This is also the cause of Oomph bug 575240.

Bug: 574806
Change-Id: I3e4ad292d1746dc54d9a83a857f6979cba9fb0d0
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>