Update patch set 2
Patch Set 2:

(2 comments)

- adding a remote once works successfully
- adding a second new remote seems impossible since the "add remote" entry in the combo disappears after adding one remote

- resizing the wizard does no longer resize the list of configured refspecs

the following features of the old wizard are lost
- fetching from a custom URI without adding a remote
- fetching from a git bundle

The button "Add all branches spec" is not disabled when a refspec for all branches is already existing. Clicking this button adds another refs/spec/*:refs/remotes/origin/* refspec and the duplicate refspecs are highlighted in red. This problem already existed before this change.

Patch-set: 2
CC: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>