Update patch set 3
Patch Set 3:

> - adding a remote once works successfully
> - adding a second new remote seems impossible since the "add remote" entry in the combo disappears after adding one remote

This should be fixed with latest patchset for https://git.eclipse.org/r/c/egit/egit/+/184360

> - resizing the wizard does no longer resize the list of configured refspecs

This should not be fixed with the latest patch set here.

> the following features of the old wizard are lost
> - fetching from a custom URI without adding a remote

Is this a features we need to keep? While it can be convenient from time to time, I didn't miss it when it was removed from the "Push" dialog. I actually even have the impression that always adding a remote for a custom URI is actually a good practice that is fine to be encouraged.
People who don't want the remote to be persisted can still remove it later, when/if it annoys them.

> - fetching from a git bundle

I was not aware of this functionality. Is it specific to Fetch, or is a "git bundle" somehow equivalent to (and covered as) a read-only remote?

Patch-set: 3