Update patch set 7
Patch Set 7:

The behavior we're seeing with "wrong remotes" is the behavior that would happen if first creating a remote in the deciated wizard, and then selecting the remote in the push wizard.
This issue is not really about the combo IMO, it's more that checking the remote should happen during the remote creation wizard, so it seems more like a topic for a separate change.

Patch-set: 7