Don't refresh the CommitEditor when not visible
A CommitEditor refreshes when Refs change in the repository containing
the commit shown (branches or tags might have changed). Let's do that
only when the editor and in particular its commit page is visible.

Otherwise all CommitEditors open on commits from the repository will
suddenly start refreshing and contend for reading information from the
repository and then doing UI updates, even if not visible.

Change-Id: I7052a430ac4c04c3ff54dc325e7fb62d6015f350
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>