[refresh] Ignore paths in closed projects
With nested closed projects, it is otherwise possible that a folder in
an open project higher up in the nesting hierarchy would be marked as
to be fully refreshed, which might then cause changed files in other
nested projects not to be refreshed, depending on resource filters in
the upper project. If that folder containing the nested projects is
filtered from the resource tree, refreshing it would do nothing.

Bug: 575720
Change-Id: I9331f75c6f88786043502d93969f3378502e122b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>