Handle long error messages in InputDialog
Increase the height of the dialog if it isn't sufficient to show the
whole error message. Do not make the dialog resizeable by the user;
if the user _does_ resize the dialog, the layout is ugly: the message
text above the input field grabs too much space.

The dialog height is only increased, but never decreased, as that would
lead to a constantly changing dialog when the user types and the
validator returns messages of different lengths.

Bug: 575695
Change-Id: I383090e9ce1b27dcb63d15b4c73beb17dd5d1b3f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>