[releng] Switch to BREE Java 11
Update the BREE in all MANIFEST.MF and .classpaths to JavaSE-11. Update
root pom.xml to use --release 11, and to compile against the egit-4.17
target by default. Update the JDT compiler settings.

Update tycho to 2.4.0 and the JBoss tycho tools to 2.2.0. Require maven
3.6.3. (3.8.2 is not installed on CI yet?) Remove pack200.

Bump spotbugs to 4.4.1; the previously used 3.1.2 uses a groovy version
that performs illegal reflective accesses that are forbidden in later
Java versions.

egit-4.17 corresponds to Eclipse 2020-09, which is the first Eclipse
version that required Java 11, and which is EGit's new baseline as of
EGit 6.0.

In the Oomph setup, use egit-4.18 as default. Eclipse 2020-09 had a
critical bug on OS X that shows up in a child Eclipse even on OS X
10.14.6 and that breaks table rendering with custom painting (bug
565969), which breaks the rendering of the EGit history view.

Hence developers are encouraged to use a target platform newer than
egit-4.17. The CI build will still compile against that platform to
ensure that EGit works on its baseline version.

Adapt tests so that they work with either target platform. Eclipse
platform changed the message in a TitleAreaDialog from Text to Label
in 4.18 (bug 564535) and changed it back to Text in 4.21 (bug 574071).
This makes SWTBot tests for this message dependent on the JFace version,
and causes particular trouble in WizardDialogs.

Make TestUtil.processUIEvents() a no-op: running event loops in tests
may block the tests if there is an asyncExec queued that opens a modal
dialog. Adapt tests not to rely on UI event processing.

Fix one bug in GenerateHistoryJob which sometimes didn't update the
history view if there were no commits found at all.

Add a null check in SpellCheckableMessageArea to work around an NPE
occurring on gtk, where apparently forceFocus() may get called on a
disposed widget.

Bug: 576238
Change-Id: I50b3d6d912e5cce2b5f9fd907c8c66cbde70be73
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>