Update patch set 5
Patch Set 5:

This is getting tedious. First hours wasted because of that stupid Text->Label->Text change in TitleAreaDialog (should I send a bill to karankun?), and when I finally have a build that runs perfectly fine locally on OS X, it fails on CI, first because CI doesn't have the newest maven, then because on gtk somehow SpellcheckableMessageArea.forceFocus() is called even though it's disposed (doesn't happen on OS X, and didn't happen on Java 8 with the egit-4.6 target), and then it just hangs in BranchAndResetActionTest.

Patch-set: 5