Update patch set 6
Patch Set 6:

The blocking in BranchAndResetTest is indeed a bug in the tests. The test performs a TestUtil.processUIEvents(), which gets run in a Display.syncExec(). One of the processed events opens a modal dialog, which then runs its own event loop. So the test continues only once that modal dialog is closed. But nothing closes it; it would be closed by the test later. But the test can't continue since it's stuck in the dialog's event loop.

Not sure yet how to fix this. Looks nasty. (And I have no idea why it doesn't happen with Java 8 egit-4.6, nor on OS X.)

Patch-set: 6