Use a lambda for the RememberHostTask
This task remembers host names of hosts the user has successfully
cloned a repository from. This is used to determine whether an
arbitrary http(s) URI might be a git URI: if it doesn't end in
".git", it's considered a potential git URI only if the host matches
one of the "known hosts".

Change-Id: I3b2ead9a847cb7bab3b3e274cd0ba19db3b7b295
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>